+++ design-system/logo/src/jira-product-discovery-logo/icon.tsx
@@ -32,7 +32,6 @@ const svg = ({
 
   return `
   <svg
-  width="32"
   height="32"
   viewBox="0 0 32 32"
   fill="none"