+++ design-system/logo/src/jira-product-discovery-logo/logo.tsx
@@ -33,7 +33,6 @@ const svg = ({
 
   return `
   <svg
-    width="264"
     height="32"
     viewBox="0 0 264 32"
     fill="none"