+++ editor/editor-plugin-table/src/__tests__/visual-regression/sticky-header.ts
@@ -38,7 +38,8 @@ describe('Snapshot Test: Table', () => {
   });
 
   describe('sticky header', () => {
-    it('should align with table cell when active', async () => {
+    // FIXME: This test was automatically skipped due to failure on 14/02/2023: https://product-fabric.atlassian.net/browse/ED-16900
+    it.skip('should align with table cell when active', async () => {
       await initEditor(page, stickyHeaderWithHorizontalScroll);
 
       await clickFirstCell(page, true);